backend/auth/services/security.py
import os
from datetime import datetime, timedelta, timezone
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import HTTPException, status
from pydantic import ValidationError
from typing import Optional, Dict, List
from ..config import security as settings
from ..schemas.token import TokenData
import secrets
import base64

class SecurityService:
    def __init__(self):
        self.pwd_context = CryptContext(
            schemes=["argon2", "bcrypt"],
            deprecated="auto",
            argon2__memory_cost=65536,
            argon2__parallelism=8
        )
        self.token_blacklist = set()

    def verify_password(self, plain_password: str, hashed_password: str) -> bool:
        """Secure password verification with Argon2 and fallback"""
        # --- Start of Masked Code ---
        # Password verification logic removed for submission.
        # This typically involves using self.pwd_context.verify().
        # --- End of Masked Code ---
        # Placeholder return for demonstration
        # Returning False as a safe default, actual logic is masked.
        return False

    def get_password_hash(self, password: str) -> str:
        """Generate secure password hash using Argon2"""
        # --- Start of Masked Code ---
        # Password hashing logic removed for submission.
        # This typically involves using self.pwd_context.hash().
        # --- End of Masked Code ---
        # Placeholder return for demonstration
        return "masked_hashed_password"

    def validate_password_complexity(self, password: str) -> None:
        """Enforce password policy requirements"""
        # --- Start of Masked Code ---
        # Password complexity validation logic removed for submission.
        # This typically involves checking length, character types (upper, lower, digit, special)
        # and raising an HTTPException if criteria are not met.
        # --- End of Masked Code ---
        # Placeholder for demonstration
        pass # Assuming validation passes for masked version

    def create_access_token(
        self, 
        subject: str, 
        scopes: List[str],
        role_id: int,
        org_id: int,
        token_expiry: int,
        expires_delta: Optional[timedelta] = None,
    ) -> str:
        """Generate JWT with security claims including role_id, org_id, and a unique jti."""
        # --- Start of Masked Code ---
        # JWT creation logic removed for submission.
        # This typically involves:
        # 1. Calculating expiry time.
        # 2. Generating a unique token ID (jti).
        # 3. Defining the payload dictionary with claims (sub, scopes, role_id, org_id, iss, aud, iat, exp, jti).
        # 4. Encoding the payload into a JWT string using jose.jwt.encode().
        # --- End of Masked Code ---
        # Placeholder return for demonstration
        return "masked_jwt_token"

    def decode_token(self, token: str) -> Dict:
        """Validate JWT with full claim verification, check blacklisting, issuer, audience, etc."""
        # --- Start of Masked Code ---
        # JWT decoding and validation logic removed for submission.
        # This typically involves:
        # 1. Using jose.jwt.decode() with secret key, algorithms, audience, and issuer.
        # 2. Checking if the token's jti is in the blacklist.
        # 3. Handling JWTError and raising appropriate HTTPExceptions.
        # --- End of Masked Code ---
        # Placeholder return/raise for demonstration
        # Returning a placeholder dict, actual logic might raise HTTPException
        return {"sub": "masked_user", "jti": "masked_jti", "scopes": [], "role_id": 0, "org_id": 0}

    def invalidate_token(self, token: str) -> None:
        """Add token to revocation list"""
        # --- Start of Masked Code ---
        # Token invalidation logic removed for submission.
        # This typically involves decoding the token to get the jti and adding it to self.token_blacklist.
        # --- End of Masked Code ---
        # Placeholder for demonstration
        pass

    def generate_jti(self) -> str:
        """Create unique JWT identifier for revocation tracking"""
        # --- Start of Masked Code ---
        # JTI generation logic removed for submission.
        # This typically involves generating random bytes and encoding them.
        # --- End of Masked Code ---
        # Placeholder return for demonstration
        return "masked_unique_jti"
