backend/chatbot_service/services/general_chat.py
from llama_index.core.llms import ChatMessage, MessageRole
from .utils import read_prompt
from langfuse.decorators import observe, langfuse_context
from .model_pricing import ModelRegistry
import time

"""
General Chat Handler module for the GRAID application.
This module provides functionality for handling general conversational
interactions that don't require content retrieval from the video data.

Implementation details are masked for security and intellectual property protection.
"""

class LLMGeneralChat:
    """
    Handles general conversational interactions using a large language model.
    
    This class manages general chat capabilities, including:
    - Loading and formatting chat prompts
    - Managing conversation history context
    - Sending formatted prompts to the LLM
    - Tracking token usage and costs via Langfuse
    
    Implementation details masked for security and intellectual property protection.
    """
    
    def __init__(self, client, chat_prompt_file):
        """
        Initialize the general chat handler.
        
        Args:
            client: The LLM client to use for chat completions
            chat_prompt_file: Path to the YAML file containing the chat prompt template
            
        Implementation details masked for security and intellectual property protection.
        """
        # Implementation masked for security
        pass

    @observe(as_type="generation",name = "GeneralChat_Intent")
    async def run(self, query: str, video_description, chat_history: list = None, model_name: str = 'meta-llama/llama-4-scout-17b-16e-instruct'):
        """
        Process a general chat query and generate a response.
        
        This function:
        1. Loads the prompt template from the specified file
        2. Formats the prompt with chat history and video context
        3. Sends the formatted prompt to the LLM
        4. Calculates token usage and costs
        5. Updates Langfuse with tracking information
        6. Returns the generated response
        
        Args:
            query (str): The user's chat query
            video_description: Description of the current video for context
            chat_history (list, optional): Previous chat messages for context
            model_name (str): The LLM model to use for generation
            
        Returns:
            str: The generated chat response
            
        Implementation details masked for security and intellectual property protection.
        """
        # Implementation masked for security
        pass