backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "vrm-hub",email = "vinay97.pro@gmail.com"}
]
readme = "README.md"
requires-python = "<3.13,>=3.9"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "google-cloud-storage (>=3.0.0,<4.0.0)",
    "sqlalchemy-orm (>=1.2.10,<2.0.0)",
    "llama-index-core (==0.12.14)",
    "llama-index-embeddings-huggingface (==0.5.0)",
    "llama-index-llms-groq (==0.3.1)",
    "pandas (==2.2.1)",
    "llama-index-vector-stores-pinecone (==0.4.4)",
    "pinecone-client (==5.0.1)",
    "google-cloud-secret-manager (==2.13.0)",
    "google-resumable-media (>=2.7.2,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langsmith (>=0.3.13,<0.4.0)",
    "langfuse (>=2.59.7,<3.0.0)",
    "wsproto (>=1.2.0,<2.0.0)",
    "llama-index-embeddings-openai (>=0.3.1,<0.4.0)",
    "openai (>=1.68.2,<2.0.0)",
    "groq (>=0.22.0,<0.23.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pre-commit = "^4.1.0"
